+++ tests/test_elfs/overflow.rs
@@ -10,8 +10,12 @@ fn main() -> ! {
 }
 
 fn ack(m: u32, n: u32) -> u32 {
-    let array = [0u8; 32 * 1024];
-    defmt::info!("ack(m={}, n={}, SP={})", m, n, array.as_ptr());
+    // waste stack space to trigger a stack overflow
+    let mut buffer = [0u8; 16 * 1024];
+    // estimate of the Stack Pointer register
+    let sp = buffer.as_mut_ptr();
+    defmt::println!("ack(m={=u32}, n={=u32}, SP={:x})", m, n, sp);
+
     if m == 0 {
         n + 1
     } else {