+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+- [#385] Update snapshot tests and test stack canary
 - [#384] Satisfy clippy
 - [#383] Fix unwinding
 - [#381] Add feature `ftdi` to enable `ftdi` debuggers
@@ -13,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - [#379] Update to `probe-rs v0.17`
 - [#378] Update to `probe-rs v0.16`
 
+[#385]: https://github.com/knurling-rs/probe-run/pull/385
 [#384]: https://github.com/knurling-rs/probe-run/pull/384
 [#383]: https://github.com/knurling-rs/probe-run/pull/383
 [#381]: https://github.com/knurling-rs/probe-run/pull/381